Add upload method
