--- src/uploadImages/typing.d.ts
@@ -1,4 +1,0 @@
-declare module '*.json' {
-  const value: any;
-  export default value;
-}
